Refactor cockpit markdown renderer
